requirements.txt
djangorestframework
django-reversion
